Let default config be easily overridable
If a configuration file is added with `phpenv config-add custum.ini`,
and the config tries to override a default value of `travis.ini`, it
will lose, as `travis.ini` is read later if the file starts with a
letter earlier in the alphabet than t.

This PR make the default easily overridable.

See https://travis-ci.org/joepvd/travis-experiment/builds/281200514 for
details.